Replace include by include_tasks
Starting with ansible 2.4, when using include we get:

The use of 'include' for tasks has been deprecated.

Besides, there is a weird error at the openstack gates in which tasks
are not executed in the expected order and I was told that probably
the include syntax is breaking things as the ODL role is called using
the new include_role syntax.

Change-Id: I4faef6648543f70021b7b159ecd355236876bfe7
Signed-off-by: Manuel Buil <mbuil@suse.com>