Modify rpm_repo_url condition
So far, if odl_version was not defined, we would use the latest release.
Unfortunately, ansible 'lookup' will return an empty string if the
variable does not exist and thus odl_version will always be defined:

https://github.com/ansible/ansible/issues/17329

We use ansible 'lookup' in xci to collect the values from env variables

To overcome this, we will allow a new value for odl_version in XCI:
"latest_release". This patch adapts to that new possible value

Change-Id: Ie4d76ed576ddeded130a863f847cd51cb7491456
Signed-off-by: Manuel Buil <mbuil@suse.com>