Add tests for odl user/password feature
The new feature allows creating/deleting ODL users. This patchset
introduces tests to ensure everything works as expected. The tests
consists of creating and deleting the users. The verification is
done via running the CLI command and saving the output in variable.
Assertions are made whether on these variables, if the desired user
is present or absent in them.

This change also requires a new job to be added in the RelEng project.
The required RelEng patchset is:
Ib950fbb1f8952c794380af03c25f9f4987c65b73

Note: Originally cross verification via API calls was implemented.
But had to be dropped because ODL was not listening on the desired
port. This needs further debugging.

Change-Id: I702245e1456e12c5a64b4c8fe1810101edb5abce
Signed-off-by: Taseer Ahmed <taseer94@gmail.com>