Add ODL log configuration
JIRA: INTPAK-40

Parallels the Puppet logic. Initially log levels is an empty
dictionary of the 'logger name' and 'log level'. The user can
set the logger name and log level by their choice.
Ansible implements a similar approach. By default, log_levels
is empty, but can be over written in the vars.

This patch also sets the logging mechanism, either file or console.
Default mechanism is set to console. In case of file, the log
rotation strategy is also configured.

Change-Id: Ifa39927bc77b47490d3e1b56f23aa359316bc3bf
Signed-off-by: Taseer Ahmed <taseer94@gmail.com>