Add tests for ODL log configuration
By default, the log-level for console should be info. The relevant
job overrides this to DEBUG, and these set of assertion verify the
desired behaviour, i.e the log-level is set to DEBUG.

Since, it is only possible to have one log mode, console or file,
this patch only tests the 'console' mode.

JIRA: INTPAK-40
Change-Id: I9b0ec79e41c25a4eeeef14f0d89572907b265281
Signed-off-by: Taseer Ahmed <taseer94@gmail.com>