Increase API retries and timeout
The jobs are failing because API is not responding within the specified
time of 5 seconds with 5 retries. This patch increases each to 20.

This patch also adds an extra task of waiting for the API to respond
before moving to the rest of the tasks. This allows enough waiting time
for ODL to be fully functional.

Change-Id: I852660abf34d91fcd547767b5249689b415ca94b
Signed-off-by: Taseer <taseer94@gmail.com>